+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * Show the addon status on `scalingo addons` [#604](https://github.com/Scalingo/cli/pull/604)
 * Add informative error in case of container type error when scaling an application
   [602](https://github.com/Scalingo/cli/pull/602)
+* `ps` command is renamed `scale` [#631](https://github.com/Scalingo/cli/pull/631)
 
 ### 1.19.3